request: log HTTP requests
